feat(api): build out well state for liquid height tracking (#15681)
Overview

re EXEC-603

This PR adds a new WellState class. This class currently has only one attribute, measured_liquid_heights, which is used in liquid height tracking. We created this new class(rather than just adding this attribute to labware state or something related), to allow for more detailed well state tracking in the future.
Details

measured_liquid_heights contains a Dict[labware_id, Dict[well_name, LiquidHeightInfo]]. This dictionary structure is meant to give time-efficient access to all wells of a particular labware, since the dictionary is primarily indexed by labware_id.

The presence/non-presence of a well in the WellState class indicates the following:

    If a well is not present: that well has never been probed.
    If a well is present but has a height value of 0: the most recent probe into that well found no liquid.
    If a well is present and has a height value of > 0: the most recent probe into that well found liquid.

A quick word on the implementation of the new LiquidHeightInfo type. It has two attributes, height and last_measured. The last_measured attribute currently stores as a datetime, but it could be useful in the future to consider other alternatives. Maybe protocol steps?(or something similar). The reason we have a time-related attribute at all is because right now we don't track things like aspirates and dispenses, or even user inputted liquid loading(if that's the correct term). All liquid height state is determined only by doing a liquid_probe, so it's useful to know what the last time was we did a liquid_probe on this well.

For the purposes of the StateSummary, a new type was defined called LiquidHeightSummary. It has identifying information about the labware/well and the contents of the LiquidHeightInfo type.
Test Plan

Unit tests written and passed for well_view and well_store.
Changelog

    Created WellState class, along with complementary WellView and WellStore
    Created LiquidHeightInfo type, which is used to store info about recent liquid height measurements.
    Created LiquidHeightSummary type, which is used in StateSummary.

Review requests
Risk assessment

Low.