fix(app): liquid state needs analysis
We were defaulting the liquids-confirmed state to true if there weren't
any liquids, but we can only know there aren't any liquids if we've
actually received an analysis that says that - until then, we shouldn't
show anything at all. Make the liquids-confirmed state always be false
until we get an analysis, and only _then_ set it to true if we don't
have any.

Closes RQA-3157