fix(app): fix undismissable calibration modals after e-stop event (#16216)
Closes RQA-3155

If you throw an estop event during pipette or gripper calibration flows, the "exit" button in the modal header does not properly fire the onClose event, making it impossible to exit the modal. Currently, onClose requires the home command to succeed in order to close the modal. Instead, let's just make the modal close regardless of whether the home command succeeds or fails.

Note that module calibration flows already have this logic, it's just these two flows that don't.