fix(robot-server): matching RTPs check during analysis of non-RTP protocols  (#16171)
Closes RQA-3118

# Overview

The check for matching RTPs had a bug where if a protocol had no RTPs,
it would deem that protocol to have 'no previously matching analyses'
and hence would always re-analyze the protocol, even when the protocol
has not changed.

This PR fixes that check and adds tests so that we'll know if it happens
again.

Fixing this bug revealed another bug that the legacy context plugin task
being created while setting up the protocol runner, was never being
stopped/ cancelled, hence preventing the server from shutting down. To
fix that, I am stopping the orchestrator when the analyzer's destructor
is called.

## Test Plan and Hands on Testing

Follow the instructions in the RQA ticket above and see that the
erroneous behavior is not seen.

## Review requests

- check that the tests cover all cases of analyzing RTP & non-RTP
protocols with various combinations of new and previous parameters (if
any)

## Risk assessment

Medium. Bug fix, but need to make sure that all combinations of possible
protocol analysis params are covered. Also affects server shutdown.

---------

Co-authored-by: Josh McVey <josh.mcvey@opentrons.com>