fix(app): display only one conflict with thermocycler in run setup
there are several combinations of labware, fixtures, and modules in A1 and B1 that generate two
conflicts with a configured thermocycler. this PR filters out one of 2 conflicts from display when
both conflicts are with single slot labware in a protocol. other scenarios should be covered by the
modules or fixtures lists.

closes RQA-3112