refactor(app): sort pages by app (#16227)
We make two apps. They do things that are very similar in gross
structure but differ quite a lot in fine detail, and have very similar
names. This is infinitely frustrating. To make it finitely frustrating,
or at least a lower cardinality of infinitely frustrating, let's do some
~*~sorting~*~

Pages are a good place to start, because they're always used in either
the desktop or the ODD, never both. Take the time to sort them into
subdirectories and update import paths. There are no code changes
outside import statements or mock calls in this PR.

The next step is probably a PR to make sure nothing from `organisms/` is
importing from `pages/`, then maybe some kind of sorting applied to
organisms.