refactor(app): Add hook for on-the-fly maintenance run commands (#16249)
Closes EXEC-699

In pre-release work, we decided to add a drop tip CTA option of "skip and home pipettes". In order to do this, we needed to create a maintenance run, initiate commands, and then close the maintenance context. There wasn't any general purpose way of doing this (we always had a wizard flow when issuing commands up to this point), so the solution was to commandeer existing drop tip wizard hooks to do this for us.

While that works, it has a few issues:

* It was pretty hackily implemented, and it kind of had to be without a dedicated hook.
* It makes following drop tip wizard more difficult to follow.
* If we every want to issue commands to a robot outside of a wizard flow, we have to add to the tech debt.
Since 8.1 should entail some drop tip wizard refactoring, this seems like a pretty good thing to refactor.

This adds a new useRobotControlCommands, which is a refactor of what the more specific drop tip wizard useDropTipMaintenanceRun. useDropTipMaintenanceRun was half doing this functionality plus other things (useDropTipMaintenanceRun now only does the other things!). This PR cleans up a lot of the cruft/paves the way for more drop tip refactors in 8.1 that were a result of the refactor.