refactor(api): Move pipette location updates into command implementations (#16160)
## Overview

This accomplishes a chunk of EXEC-652. It also sets the pattern for
EXEC-639 in general, at least to start with.

## Test Plan and Hands on Testing

* [x] Run some protocols in general on a real robot and make sure
nothing weird happens with path planning.
* [x] Check analysis snapshot tests to make sure there are no changes.

## Changelog

* When a command succeeds, or fails with a defined error, it now passes
a new piece of data to engine internals: the `state_update`.

The `state_update` lets you express common things that commands do, like
"set pipette `abc123`'s logical location to well `A1` in labware
`def456`". The `StateStore`s listen for `state_updates` and perform the
actual modifications.

There are some convenience methods to help command implementations build
up a `StateUpdate` step by step, as they physically affect the robot in
different ways. This is done in a way that retains the fundamental
behavior of Protocol Engine where state is only changed in between
commands, not during them.
  
* Partially port a handful of commands. They now use `state_update` to
update the pipette location, instead of `PipetteStore` doing stuff like
`isinstance(command, MoveToWell)`.

The follow-up work is to keep going with more commands, and more kinds
of state, like liquid amounts.

## Risk assessment

High. This is intrusive to how protocols execute, e.g. it affects path
planning. It's difficult to test that the overall integrated behavior is
preserved. G-code tests cover this a bit, but I don't think those have
kept up with new commands.