fix(api): use critical point instead of primary nozzle when doing deck conflict check (#16268)
Closes RQA-3175

# Overview

There was a bug in the deck conflict checker that it wasn't considering
the change in critical point of pipette in use when addressing
reservoirs. This was leading to incorrect conflict checks when moving to
any labware that had the `centerMultichannelOnWells` quirk.

This PR fixes that by correctly finding the pipette's boundaries when
its *critical point* is moved to the destination in question rather than
its primary nozzle (the default critical point).

## Test Plan and Hands on Testing

- Added integration tests that uses reservoirs with row and channel
configurations
- Existing unit and integration tests should remain unaffected
- [x] Add tests for conflict checks with 8-channel pipette partial
column configuration
- [x] Test on a robot that existing pipette movements with different
pipettes in different configurations is not affected

## Review requests

- Make sure the logic checks out & test on robot

## Risk assessment

Low. We have quite good test coverage for a lot of cases so any
unexpected changes should get caught in the tests easily.