refactor(api): Delete `DefinedErrorData.private` (#16262)
## Overview

A trivial refactor as part of EXEC-652.

## Changelog

The `DefinedErrorData.private` attribute was obsoleted by PR #16160's
`StateUpdate` mechanism. PR #16261 deleted the last use of it, so we can
delete the attribute now.

`DefinedErrorData`'s sibling class, `SuccessData`, also has an identical
`private` attribute. We're not quite ready to delete that one, though.