refactor(app): allow "absolute" imports in app js (#16288)
This PR adds configuration to tsc for typechecking and lint, vite for
bundling, and vitest for testing to allow code in `app/` to import
things via paths that look absolute from the root of the repo (but are
really just a specific binding). Any path in an import or mock statement
that starts with `/app` will get rewritten to (the appropriate resolved
absolute path for) `app/src/`.

In addition, rewrite any import that was a relative path that used a
string of `../` to get all the way to the top of the repo before going
down to another top-level directory as such an absolute-path import.
There are a lot of these, which is kind of gross. Each commit after the
first handles this rewriting for another kind of import target.

## Note on merging and reviewing

This PR will conflict with basically any change to the app code, so I'm
not going to make it always up to date. Rather, please review this as
is, including checking it out and building/bundling the app if you want,
and when it's approved I'll do all the merge conflicts then.

## review requests
- Is this what we want to do?