refactor(robot-server): Clean up global resources in FILO order (#16286)
## Overview

This refactors the way the server sets up and tears down its global
resources. This may help with concurrency problems like the deadlock
found in #16171/RQA-3118, and I think it's a good idea for other
reasons, too.

## Details

The server has many global singleton resources that need setup and
cleanup—`ProtocolStore`, `HardwareControlAPI`, `NotificationClient`, and
so on.

Historically, we've done that in a pair of `startup()`/`shutdown()`
functions, because that's what FastAPI encouraged. But modern FastAPI
versions have a better way of doing it, with a single "lifespan" context
manager. The advantages of the new way are:

* It more strongly enforces that things are set up and torn down in FILO
order. This may help with concurrency problems like the deadlock found
in #16171.
* It encourages us to write the underlying resources as context
managers, which, again, will help avoid concurrency problems and
resource leaks.
* It will help undo some of the global state and callback soup that we
have going on in this area.

For example, formerly, when we needed to pass something between startup
and shutdown, we had to store it on the global `AppState`. Now, it can
just be a local variable inside the lifespan context manager.

## Changelog

* Port the top-level `startup()` and `shutdown()` functions to a single
context manager.
* Port some, but not all, of the underlying resources to context
managers. Future PRs can do the same thing for the remaining ones.