refactor(app): move ProtocolUpload hooks, start moving Devices hooks (#16324)
The original aim of this PR was to untangle `app/src/organisms/Devices`
from everything, particularly in the ODD, that imports it, but it got
way too big so this is sort of the first half. It
- Moves the remaining run control hooks from `ProtocolUpload` to the run
resources
- Removes some last pages imports from organisms (these had to do with
OT-2 calibration)
- Move a first pass of analytics hooks from devices to `redux-resources`
- Move some of the key redux interface hooks (`useIsRobot`, `useIsFlex`,
dependencies, friends) to `redux-resources`

There are no code changes other than imports, mock statements, and new
index files.