feat(api, robot-server, shared-data): Add single and multi wavelength read capabilities to the Plate Reader. (#16273)
The Byonoy Absorbance 96 Plate Reader can initialize in either `single` or `multi` measure modes, depending on the mode the device can read a single or multiple wavelengths at once. In `single` mode the device can optionally take a `reference wavelength` as a parameter, so the returned data is processed automatically on the device. This functionality was always available in the Byonoy library, we now expose that API through the Opentrons hardware controller and Protocol Engine to be used in a protocol.