Flip "enable error recovery" setting to "disable error recovery."
Assuming we use the /settings HTTP endpoints for this, this matches the existing convention of the default, `null`, being equivalent to `false`. That convention is sort of built into the HTTP API: there is otherwise no way for a client to tell whether a value of `null` means the UI switch should be rendered as engaged or disengaged.