refactor(app): remove all Devices references from ODD contexts (#16347)
This moves the rest of the hooks and UI elements that were stored in
Devices/ out into their own resources, molecules, or organisms as the
case may be. After this, there are no references to Devices/ inside any
ts(x) file in `app/src/organisms/ODD` or `app/src/pages/ODD`.

As with previous PRs of this type, no functional changes - the only
changes are in the location of files, import paths, and mock paths.