chore: adapt react imports to new jsx transform (#16353)
In react 17, various kinds of tooling added a new step to transform JSX
code to JS code that didn't require importing react. That means we can
turn off the very slow eslint step that checks this, if we transform our
code to the new style by removing many react imports.


https://legacy.reactjs.org/blog/2020/09/22/introducing-the-new-jsx-transform.html#removing-unused-react-imports

This PR does that. It touches a ton of files and I don't think reading
through it will help; spotcheck it, run dev code, etc. I think it's
nice!