fix: make relative links absolute in nav for docs subdomain (#16360)
# Overview

Our shared nav resources have a mix of relative and absolute links. This
is not great!

It means we have to edit the actual HTML for inclusion on the docs site.
This PR does a find-and-replace to accomplish that.

## Test Plan and Hands on Testing

[Sandbox](http://sandbox.docs.opentrons.com/docs-fix-nav-links/v2/)
I clicked on all the links this time.

## Changelog

Replace `href="/` with `href="https://opentrons.com/` across the board.

## Review requests

Any sneaky other link format I could have missed?

## Risk assessment

low. we will have to replicate this in the future if we have to update
the nav (and its new version doesn't fix this problem)