refactor(app): move all desktop-only organisms to Desktop (#16369)
Finishes the desktop side of this big reorganization of app code. At
this point, any organism that is used only in the Desktop app now lives
in `app/organisms/Desktop`. There are no imports up the tree of life
(i.e. importing organisms from molecules or pages from organisms); there
are no imports from shared or ODD code into `Desktop` at organism or
page level.

The next step is to do the same for ODD. This should be a lot easier
since a lot of the things that tangled everything up are now resolved.