chore(app): add eslint custom rules for imports
Add an eslint custom rule for imports "up the tree of life" - no
importing organisms from molecules and molecules from atoms and so on.

Right now, this is confined to the app because that's the only domain
that we've fixed up, but we can change that in eslintrc whenever we're
confident.

Also, add rules for importing ODD from Desktop and Desktop from ODD and
either from shared and any UI from any structure.