chore(app): add custom eslint rules for imports (#16373)
Now that we've gone to all this trouble to separate app code so that
nothing imports from higher on the tree of life and desktop and ODD are
cleanly separated (well, ODD isn't yet, but it soon will be) it's Quite
Important To My Emotional Health And State Of Mind that it stays that
way.

A good way to make it stay that way is a custom eslint hook, which
was... fun... to write, but now can be a place we add more stuff like
this if we want in the future.

That plugin lives as a module that's added as a dev-dependency to the
root package json (because eslint <9 needs plugins to work this way,
sadly) so you'll have to run setup-js, but then it will yell at you if
you import up the tree of life or across applications in `app/` (it can
also do it elsewhere but it's disabled there - normal eslint rule
overrides can change this if you want).

Oh and also there were a couple lint fixes. Nothing big, just like 100
files of random little imports that violated the rules that I missed
over the previous PRs.