feat(app): Add more location options to drop tip flows (#16367)
Closes EXEC-720

This commit adds predefined options for performing drop tip operations based on current deck configuration. This commit does not add support for "returning to labware".

Unlike the current wizard's "choose location" behavior, when moving to a waste chute or trash bin, we have the ability to move the pipette in the Z-axis without risking any sort of collision, assuming the pipette with a tip attached doesn't collide into the waste chute/trash bin. Because we can't know the length of the tips on the pipette during an out-of-run drop tip flow, we take the conservative approach and assume a 1000ul tip is attached.

When moving to a trash bin/waste chute, the robot always homes the z-axis (not the plungers), does the appropriate moveToAddressableArea, and then the drop tip/blowout aciton.

Note that OT-2s currently only have one option, "Choose deck location", but still see the option selection screen. However, once "return to labware" support is added, this will change.