refactor(app): move ODD organisms into organisms/ODD (#16376)
There are a couple weird ones that are technically ODD-only but are
imported from some other organisms that provide them if they're
targeting ODD and not if they're targeting Desktop - the instrument
wizards will pop a firmware update takeover and use the firmware update
modal as an interstitial on ODD. Not sure what to do with those, left
them and their dependents in the same place.

This is the final PR in this series of code reorganizations! Hooray!