feat(app): add localization feature flag (#16341)
Adds the localization feature flag to put localization dev work behind,
the initial zh translation assets, and temp language toggles to enable
dev work. For now, when the feature flag is enabled, the desktop
language toggle is in the app settings advanced tab, ODD in the robot
settings page. These will be removed when designs are implemented.

closes PLAT-496, re PLAT-493