feat(app): hide door open modal in intervention
We display an InterventionModal when a protocol has an action that needs
a human to help. That's a full-screen takeover on both ODD and Desktop.
The problem is that most of the time they'll need to open the door...
which is also a full-screen takeover on ODD and desktop, and then they
can't see the instructions they wrote themselves.

The thing to do here is to instead convey "door is open" by blocking the
continue button, which we'll do with an inline notification.

Closes RSQ-133